Update the Category Name to reflect the grouping of the controls
This may be overkill to display controls in each assembly under its own title, but, this way, we know when the controls go missing because of a bad edit in the manifest.
Also, This should look nice with controls grouped by a certain keyword instead of being thrown into a single list having unrelated controls together.